Move proposal creation / voted event listeners to own hook
