+++ src/hooks/DAO/useDAOController.ts
@@ -2,6 +2,7 @@ import { utils } from 'ethers';
 import { useSearchParams } from 'react-router-dom';
 import { useFractal } from '../../providers/App/AppProvider';
 import { useNetworkConfig } from '../../providers/NetworkConfig/NetworkConfigProvider';
+import { useAzoriusProposalListeners } from './loaders/governance/useAzoriusProposalListeners';
 import { useERC20Claim } from './loaders/governance/useERC20Claim';
 import { useSnapshotProposals } from './loaders/snapshot/useSnapshotProposals';
 import { useFractalFreeze } from './loaders/useFractalFreeze';
@@ -48,6 +49,7 @@ export default function useDAOController() {
   useFractalTreasury();
   useERC20Claim();
   useSnapshotProposals();
+  useAzoriusProposalListeners();
 
   return { invalidQuery, wrongNetwork, errorLoading };
 }
